+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Enhancements:
 
-- feat(observability/loggingendpointerrors): add support for pagination via 'NextLink' and 'PrevLink' ([#802](https://github.com/fastly/go-fastly/pull/802))
+- feat(observability/loggingendpointerrors): add support for pagination via 'NextFrom' and 'PrevFrom' fields extracted from Link headers ([#802](https://github.com/fastly/go-fastly/pull/802))
 
 ### Dependencies: